Revert "use arm-eabi-4.8 for kernel builds as default"
This reverts commit 9e62aa3b7f9733818785b7002adf1d75dc5dc304.

This was a hack to force 4.8 back in M because Google forced 4.9 on us
but at the time it wasn't available. I know I know, hard to believe that
Google would mess up like that but they did. Anyways, this is no longer
needed. Really it wasn't needed in N but some how found its way into it.


Change-Id: Iafc1924ac5d3fd004fabef14967e2db2d328c777